+++ lib/views/questions_view.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_ecoapp/models/article.dart';
 import 'package:flutter_ecoapp/models/question.dart';
 import 'package:flutter_ecoapp/views/style/colors.dart';
+import 'package:flutter_ecoapp/views/widgets/articles/article_cover.dart';
 import 'package:flutter_ecoapp/views/widgets/bottom_nav_bar.dart';
 import 'package:google_fonts/google_fonts.dart';
 
@@ -52,7 +53,10 @@ class QuestionsView extends StatelessWidget {
         ),
         child: Column(
           children: [
-            _QuestionsArticleCover(article: article),
+            ArticleCover(
+              article: article,
+              subtitle: '${article.questions!.length.toString()} preguntas',
+            ),
             SizedBox(height: 10.0,),
             Divider(thickness: 1,),
             _QuestionsList(article: article)
@@ -62,70 +66,6 @@ class QuestionsView extends StatelessWidget {
     );
   }
 
-}
-
-class _QuestionsArticleCover extends StatelessWidget {
-  const _QuestionsArticleCover({
-    Key? key,
-    required this.article,
-  }) : super(key: key);
-
-  final ArticleModel article;
-
-  @override
-  Widget build(BuildContext context) {
-    final img = article.photos.length > 0? Container(
-      width: 80,
-      height: 80,
-      clipBehavior: Clip.antiAlias,
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.5),
-            blurRadius: 5,
-            offset: Offset(0, 2)
-          )
-        ]
-      ),
-      child: Image(
-        image: NetworkImage(article.photos[0].photoUrl),
-        fit: BoxFit.cover,
-      ),
-    ) : Container();
-    
-    final info = Container(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.end,
-        children: [
-          Text(
-            article.title,
-            style: GoogleFonts.montserrat(
-              fontWeight: FontWeight.w500,
-            ),
-            textAlign: TextAlign.end,
-          ),
-          SizedBox(height: 10.0,),
-          Text(
-            '${article.questions!.length.toString()} preguntas',
-            style: GoogleFonts.montserrat(),
-            textAlign: TextAlign.end,
-          )
-        ],
-      ),
-    );
-
-    return Container(
-      child: Row(
-        mainAxisSize: MainAxisSize.max,
-        children: [
-          img,
-          SizedBox(width: 20.0),
-          Expanded(child: info)
-        ],
-      ),
-    );
-  }
 }
 
 class _QuestionsList extends StatelessWidget {
